Small presenter area refactor
